httputil/netrc_test.go
package httputil

import (
	"os"
	"testing"

	"github.com/ghetzel/go-stockutil/fileutil"
	"github.com/ghetzel/testify/require"
)

func TestNetrcPositive(t *testing.T) {
	NetrcFile = fileutil.MustWriteTempFile(
		"machine stock-httputil-test login hello password there\nmachine intentionally-left-blank\n",
		"test-ghetzel-go-stockutil-httputil",
	)

	defer os.Remove(NetrcFile)

	var u, p, ok = NetrcCredentials(``)

	require.False(t, ok)
	require.Empty(t, u)
	require.Empty(t, p)

	u, p, ok = NetrcCredentials(`nope`)

	require.False(t, ok)
	require.Empty(t, u)
	require.Empty(t, p)

	u, p, ok = NetrcCredentials(`intentionally-left-blank`)

	require.False(t, ok)
	require.Empty(t, u)
	require.Empty(t, p)

	u, p, ok = NetrcCredentials(`stock-httputil-test`)

	require.True(t, ok)
	require.Equal(t, `hello`, u)
	require.Equal(t, `there`, p)
}
